+++ CHANGELOG.md
@@ -7,8 +7,7 @@
 # 1.0.2
 
 ## Fixed
-- Fixed an issue where identifier in the queue processor's `channel_directory` attribute would store incorrect keys for candle data. 
->>>>>>> hotfix-1.0.2
+- Fixed an issue where identifier in the queue processor's `channel_directory` attribute would store incorrect keys for candle data.
 
 # 1.0